remove duplicate defines
